requirements.txt
chromadb>=0.4.0
openai>=1.0.0
google-generativeai>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
sqlalchemy
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.1.0
websockets>=10.0
python-multipart>=0.0.5
pydantic>=2.0.0
youtube-transcript-api>=0.6.0
pytz

